# Removed all warnings (issue : #70)