--- a/src/components/Faq/faq.js
+++ b/src/components/Faq/faq.js
@@ -2,7 +2,6 @@ import "./faq.css";
 import { Container, Row, Col} from "react-bootstrap";
 import "./questions.js";
 import ToggleableAnswer from "./questions.js";
-import fancyUnderline from "./../../Assets/fancyUnderline.png";
 import crossPattern from "./../../Assets/crossPattern.png";
 
 function Faq(){
@@ -10,7 +9,7 @@ function Faq(){
         <section>
             <Container className="faq-section" id="faq">
                 <Container className="faq-content">
-                    <img src={crossPattern} className="pattern"></img>
+                    <img src={crossPattern} className="pattern" alt="Cross"></img>
                     <h1 id="faq-title" className="title">Frequently Asked Questions</h1>
 
                     <Row className="faq-body">