Removed all Warning that where showing up on NPM start.
issue : #70