@Hiten1708 2022-07-11
Good Work! I think it's error free atleast in Vercel :P

For the record

After:

[21:27:50.924] Cloning github.com/MbrllaHacks/Website (Branch: nitin094/removed-warnings, Commit: 999176a)
[21:27:53.390] Cloning completed: 2.465s
[21:27:53.818] Installing build runtime...
[21:27:56.234] Build runtime installed: 2.415s
[21:27:56.745] Looking up build cache...
[21:28:00.502] Build cache downloaded [46.82 MB]: 3377.782ms
[21:28:00.713] Installing dependencies...
[21:28:04.034] 
[21:28:04.035] up to date in 3s
[21:28:04.035] 
[21:28:04.035] 194 packages are looking for funding
[21:28:04.036]   run `npm fund` for details
[21:28:04.317] Detected `package-lock.json` generated by npm 7+...
[21:28:04.318] Running "npm run build"
[21:28:04.711] 
[21:28:04.711] > mbrllahacks-website@0.1.0 build
[21:28:04.711] > react-scripts build
[21:28:04.712] 
[21:28:06.352] Creating an optimized production build...
[21:28:10.702] [BABEL] Note: The code generator has deoptimised the styling of /vercel/path0/src/Assets/nickumbrella.svg as it exceeds the max of 500KB.
[21:28:11.118] [BABEL] Note: The code generator has deoptimised the styling of undefined as it exceeds the max of 500KB.
[21:28:11.665] [BABEL] Note: The code generator has deoptimised the styling of /vercel/path0/src/Assets/pc2.svg as it exceeds the max of 500KB.
[21:28:11.777] [BABEL] Note: The code generator has deoptimised the styling of undefined as it exceeds the max of 500KB.
[21:28:12.906] [BABEL] Note: The code generator has deoptimised the styling of /vercel/path0/src/Assets/nick.svg as it exceeds the max of 500KB.
[21:28:14.433] [BABEL] Note: The code generator has deoptimised the styling of undefined as it exceeds the max of 500KB.
[21:28:26.996] Compiled successfully.
[21:28:26.997] 
[21:28:26.997] File sizes after gzip:
[21:28:26.997] 
[21:28:27.022]   64.24 kB  build/static/js/main.675d227b.js
[21:28:27.023]   26.41 kB  build/static/css/main.e4f81784.css
[21:28:27.023] 
[21:28:27.023] The project was built assuming it is hosted at /.
[21:28:27.023] You can control this with the homepage field in your package.json.
[21:28:27.023] 
[21:28:27.024] The build folder is ready to be deployed.
[21:28:27.024] You may serve it with a static server:
[21:28:27.024] 
[21:28:27.024]   npm install -g serve
[21:28:27.024]   serve -s build
[21:28:27.024] 
[21:28:27.025] Find out more about deployment here:
[21:28:27.025] 
[21:28:27.025]   https://cra.link/deployment
[21:28:27.025] 
[21:28:29.048] Generated build outputs:
[21:28:29.048]  - Static files: 22
[21:28:29.049]  - Serverless Functions: 0
[21:28:29.049]  - Edge Functions: 0
[21:28:29.049] Deployed outputs in 1s
[21:28:30.889] Build completed. Populating build cache...
[21:28:58.899] Uploading build cache [46.82 MB]...
[21:29:00.868] Build cache uploaded: 1.968s
[21:29:00.903] Done with "package.json"

Before:

[00:14:51.653] Cloning github.com/MbrllaHacks/Website (Branch: mint-FAQ, Commit: 5123336)
[00:14:53.967] Cloning completed: 2.314s
[00:14:54.325] Installing build runtime...
[00:14:56.343] Build runtime installed: 2.018s
[00:14:57.486] Looking up build cache...
[00:15:00.670] Build cache downloaded [46.83 MB]: 2516.255ms
[00:15:00.831] Installing dependencies...
[00:15:03.321] 
[00:15:03.321] up to date in 2s
[00:15:03.321] 
[00:15:03.322] 194 packages are looking for funding
[00:15:03.322]   run `npm fund` for details
[00:15:03.542] Detected `package-lock.json` generated by npm 7+...
[00:15:03.542] Running "npm run build"
[00:15:03.836] 
[00:15:03.836] > mbrllahacks-website@0.1.0 build
[00:15:03.836] > react-scripts build
[00:15:03.836] 
[00:15:05.098] Creating an optimized production build...
[00:15:08.551] [BABEL] Note: The code generator has deoptimised the styling of /vercel/path0/src/Assets/nickumbrella.svg as it exceeds the max of 500KB.
[00:15:08.928] [BABEL] Note: The code generator has deoptimised the styling of undefined as it exceeds the max of 500KB.
[00:15:09.401] [BABEL] Note: The code generator has deoptimised the styling of /vercel/path0/src/Assets/pc2.svg as it exceeds the max of 500KB.
[00:15:09.509] [BABEL] Note: The code generator has deoptimised the styling of undefined as it exceeds the max of 500KB.
[00:15:10.480] [BABEL] Note: The code generator has deoptimised the styling of /vercel/path0/src/Assets/nick.svg as it exceeds the max of 500KB.
[00:15:11.935] [BABEL] Note: The code generator has deoptimised the styling of undefined as it exceeds the max of 500KB.
[00:15:22.011] Compiled with warnings.
[00:15:22.012] 
[00:15:22.012] Warning
[00:15:22.012] (6:29521) autoprefixer: Replace color-adjust to print-color-adjust. The color-adjust shorthand is currently deprecated.
[00:15:22.012] 
[00:15:22.013] src/components/About/about.js
[00:15:22.013]   Line 6:30:  'Particle' is defined but never used  no-unused-vars
[00:15:22.013] 
[00:15:22.013] src/components/Contact/contact.js
[00:15:22.013]   Line 6:8:    'prin' is defined but never used                                                                           no-unused-vars
[00:15:22.013]   Line 33:17:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.013]   Line 36:17:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.014] 
[00:15:22.014] src/components/Faq/faq.js
[00:15:22.014]   Line 5:8:    'fancyUnderline' is defined but never used                                                                 no-unused-vars
[00:15:22.014]   Line 25:21:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.014] 
[00:15:22.014] src/components/Faq/questions.js
[00:15:22.014]   Line 2:10:  'Container' is defined but never used  no-unused-vars
[00:15:22.014]   Line 2:26:  'Col' is defined but never used        no-unused-vars
[00:15:22.015] 
[00:15:22.015] src/components/Home/home.js
[00:15:22.015]   Line 4:30:   'Particle' is defined but never used                                                                       no-unused-vars
[00:15:22.015]   Line 32:25:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.015] 
[00:15:22.015] src/components/Navbar/navbar.js
[00:15:22.015]   Line 14:11:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.015]   Line 44:15:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.015]   Line 47:15:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.015]   Line 51:15:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.016]   Line 55:15:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.016] 
[00:15:22.016] src/components/Sponsors/sponsors.js
[00:15:22.016]   Line 3:8:    'sponsorBlur' is defined but never used                                                                    no-unused-vars
[00:15:22.016]   Line 35:21:  img elements must have an alt prop, either with meaningful text, or an empty string for decorative images  jsx-a11y/alt-text
[00:15:22.016] 
[00:15:22.016] Search for the keywords to learn more about each warning.
[00:15:22.016] To ignore, add // eslint-disable-next-line to the line before.
[00:15:22.016] 
[00:15:22.016] File sizes after gzip:
[00:15:22.016] 
[00:15:22.034]   63.52 kB  build/static/js/main.270bd8c3.js
[00:15:22.034]   25.32 kB  build/static/css/main.3079521d.css
[00:15:22.034] 
[00:15:22.034] The project was built assuming it is hosted at /.
[00:15:22.034] You can control this with the homepage field in your package.json.
[00:15:22.034] 
[00:15:22.034] The build folder is ready to be deployed.
[00:15:22.034] You may serve it with a static server:
[00:15:22.035] 
[00:15:22.035]   npm install -g serve
[00:15:22.035]   serve -s build
[00:15:22.035] 
[00:15:22.035] Find out more about deployment here:
[00:15:22.035] 
[00:15:22.035]   https://cra.link/deployment
[00:15:22.035] 
[00:15:24.070] Generated build outputs:
[00:15:24.070]  - Static files: 19
[00:15:24.070]  - Serverless Functions: 0
[00:15:24.071]  - Edge Functions: 0
[00:15:24.071] Deployed outputs in 1s
[00:15:25.723] Build completed. Populating build cache...
[00:15:45.670] Uploading build cache [46.82 MB]...
[00:15:47.161] Build cache uploaded: 1.491s
[00:15:47.190] Done with "package.json"